simplify code

+++ desktop/applis/apps/bean/bean_games/pokemonbean/src/main/java/aiki/beans/facade/simulation/dto/EvLine.java
@@ -1,13 +1,13 @@
 package aiki.beans.facade.simulation.dto;
 
 public final class EvLine {
-    private short ev;
+    private int ev;
 
-    public short getEv() {
+    public int getEv() {
         return ev;
     }
 
-    public void setEv(short _ev) {
+    public void setEv(int _ev) {
         ev = _ev;
     }
 }
